requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
openai>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-dotenv>=1.0.0
pillow>=10.0.0
aiofiles>=23.0.0
aiosqlite>=0.19.0
httpx>=0.25.0
schedule>=1.2.0
matplotlib>=3.7.0
numpy>=1.24.0

# Local LLM Dependencies (optional)
aiohttp>=3.8.0
requests>=2.28.0
transformers>=4.21.0
torch>=1.12.0
tokenizers>=0.13.0
accelerate>=0.21.0
sentencepiece>=0.1.97
